Django-Scraping/requirements.txt
asgiref==3.2.7
autopep8==1.4.3
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Django==2.1.7
django-bootstrap3==11.0.0
django-bootstrap4==0.0.8
django-braces==1.13.0
django-crispy-forms==1.9.1
django-filter==2.3.0
djangorestframework==3.9.2
idna==2.8
misaka==2.1.1
Pillow==5.4.1
pkg-resources==0.0.0
plotly==4.8.1
postgres==2.2.2
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pycparser==2.19
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
retrying==1.3.3
six==1.12.0
soupsieve==1.9
sqlparse==0.3.1
urllib3==1.24.1
wincertstore==0.2